Refactor services to use dependency injection with interfaces
