+++ src/main/java/com/danielagapov/spawn/Controllers/User/Profile/UserInterestController.java
@@ -2,7 +2,7 @@ package com.danielagapov.spawn.Controllers.User.Profile;
 
 import com.danielagapov.spawn.DTOs.User.Profile.CreateUserInterestDTO;
 import com.danielagapov.spawn.DTOs.User.Profile.UserInterestDTO;
-import com.danielagapov.spawn.Services.UserInterestService;
+import com.danielagapov.spawn.Services.UserInterest.IUserInterestService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,10 +15,10 @@ import java.util.UUID;
 @RequestMapping("/api/users/{userId}/interests")
 public class UserInterestController {
 
-    private final UserInterestService userInterestService;
+    private final IUserInterestService userInterestService;
 
     @Autowired
-    public UserInterestController(UserInterestService userInterestService) {
+    public UserInterestController(IUserInterestService userInterestService) {
         this.userInterestService = userInterestService;
     }
 
@@ -30,7 +30,10 @@ public class UserInterestController {
 
     @PostMapping
     public ResponseEntity<UserInterestDTO> addUserInterest(
+            @PathVariable UUID userId,
             @RequestBody CreateUserInterestDTO createUserInterestDTO) {
+        // Set the userId from the path parameter
+        createUserInterestDTO.setUserId(userId);
         UserInterestDTO createdInterest = userInterestService.addUserInterest(createUserInterestDTO);
         return new ResponseEntity<>(createdInterest, HttpStatus.CREATED);
     }
